Acq: Reduce rate of interrupts by collecting whole message per DMA.
This reduces the rate of DMA interrupts by 16/[channels per ADC]
times per active ADC.

However it increases the amount of work needed to be done in the
interrupt handler by a proportional amount.  The compiler can do
a better job of otimising it this way though.

It also replaces a VLA with a fixed size array, which should be
faster too.