Add stub check to verify if channel is active
We decided to only turn on 1 LED at a time, and there will
be a best affinity photediode for each LED, although the other
photediodes will more or less perceive the light at same time.

So we only acquist data from the active channel, and drop data
from other channels.

Signed-off-by: Kejia Hu <kejia.hu@codethink.co.uk>